redesigned the tracker page to look more like FitN. created a number chooser class to handle weight and reps controls

+++ app/src/main/java/com/armpatch/android/workouttracker/adapters/TrackerSetAdapter.java
@@ -76,7 +76,7 @@ public class TrackerSetAdapter extends RecyclerView.Adapter<TrackerSetAdapter.Se
         return sets.size();
     }
 
-    void addSet(int measurement1, int measurement2) {
+    void addSet(float measurement1, float measurement2) {
         ExerciseSet set = new ExerciseSet(exerciseDate, exerciseName, measurement1, measurement2, getItemCount() + 1);
         new InsertSetTask(set).execute();
     }
